pyproject.toml
[project]
name = "zunel"
version = "0.1.5.post2"
description = "A lightweight personal AI assistant framework (Slack + OpenAI-compatible only)"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "zunel contributors"}
]
keywords = ["ai", "agent", "chatbot", "slack"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
license-files = [
    "LICENSE",
    "THIRD_PARTY_NOTICES.md",
]

dependencies = [
    "typer>=0.20.0,<1.0.0",
    "pydantic>=2.12.0,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "httpx>=0.28.0,<1.0.0",
    "loguru>=0.7.3,<1.0.0",
    "rich>=14.0.0,<15.0.0",
    "croniter>=6.0.0,<7.0.0",
    "prompt-toolkit>=3.0.50,<4.0.0",
    "questionary>=2.0.0,<3.0.0",
    "json-repair>=0.57.0,<1.0.0",
    "chardet>=3.0.2,<6.0.0",
    "openai>=2.8.0",
    "tiktoken>=0.12.0,<1.0.0",
    "jinja2>=3.1.0,<4.0.0",
    "dulwich>=0.22.0,<1.0.0",
    "pyyaml>=6.0,<7.0.0",
    "filelock>=3.25.2",
    "websockets>=16.0",
    "aiohttp>=3.13.5",
]

[project.optional-dependencies]
slack = [
    "slack-sdk>=3.39.0,<4.0.0",
    "slackify-markdown>=0.2.0,<1.0.0",
]
mcp = [
    "mcp>=1.26.0,<2.0.0",
]
docs = [
    "pypdf>=5.0.0,<6.0.0",
    "python-docx>=1.1.0,<2.0.0",
    "openpyxl>=3.1.0,<4.0.0",
    "python-pptx>=1.0.0,<2.0.0",
    "readability-lxml>=0.8.4,<1.0.0",
    "ddgs>=9.5.5,<10.0.0",
]
codex = [
    "oauth-cli-kit>=0.1.3,<1.0.0",
]
messaging = [
    "zunel[slack]",
]
all = [
    "zunel[slack]",
    "zunel[mcp]",
    "zunel[docs]",
    "zunel[codex]",
]
langsmith = [
    "langsmith>=0.1.0",
]
pdf = [
    "pymupdf>=1.25.0",
]
dev = [
    "pytest>=9.0.0,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "pytest-xdist>=3.0.0,<4.0.0",
    "ruff>=0.1.0",
    "pymupdf>=1.25.0",
]

[project.scripts]
zunel = "zunel.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
include = [
    "zunel/**/*.py",
    "zunel/templates/**/*.md",
    "zunel/skills/**/*.md",
    "zunel/skills/**/*.sh",
]

[tool.hatch.build.targets.wheel]
packages = ["zunel"]

[tool.hatch.build.targets.wheel.sources]
"zunel" = "zunel"

[tool.hatch.build.targets.sdist]
include = [
    "zunel/",
    "README.md",
    "LICENSE",
    "THIRD_PARTY_NOTICES.md",
    "pyproject.toml",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring external services (network, real APIs)",
]
addopts = "-m 'not integration' -n auto"

[tool.coverage.run]
source = ["zunel"]
omit = ["tests/*", "**/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
